adjusting folder structure to represent virtual module separation
