+++ SavingMoney/WritePlan/Model/PlanModel.swift
@@ -1,0 +1,9 @@
+//
+//  PlanModel.swift
+//  SavingMoney
+//
+//  Created by Paul Lee on 2021/12/13.
+//
+
+import Foundation
+public typealias PlanModel = String
