+++ SavingMoney/WritePlan/ViewController/WriteDownPlanViewController.swift
@@ -1,0 +1,57 @@
+//
+//  WriteDownPlanViewController.swift
+//  SavingMoney
+//
+//  Created by Paul Lee on 2021/12/12.
+//
+
+import UIKit
+
+public class WriteDownPlanViewController: UIViewController {
+    @IBOutlet public private(set) weak var nextBarBtnItem: UIBarButtonItem!
+    
+    @IBOutlet public private(set) weak var planTextField: UITextField!
+    
+    private var viewModel: WriteDownPlanViewModel?
+    
+    init?(coder: NSCoder, viewModel: WriteDownPlanViewModel) {
+        super.init(coder: coder)
+        self.viewModel = viewModel
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+    
+    public override func viewDidLoad() {
+        super.viewDidLoad()
+        bind()
+    }
+    
+    private func bind() {
+        bindNextBarBtnItem()
+        bindPlanTextField()
+    }
+    
+    private func bindNextBarBtnItem() {
+        viewModel?.onNextStateChange = { [nextBarBtnItem] readyForNextStep in
+            nextBarBtnItem?.isEnabled = readyForNextStep
+        }
+        
+        nextBarBtnItem.target = self
+        nextBarBtnItem.action = #selector(nextBarBtnItemTapped(_:))
+    }
+    
+    private func bindPlanTextField() {
+        planTextField.addTarget(self, action: #selector(planTextFieldEditingChanged(_:)), for: .editingChanged)
+    }
+    
+    @objc private func nextBarBtnItemTapped(_ sender: Any) {
+        viewModel?.nextStep()
+    }
+    
+    @objc private func planTextFieldEditingChanged(_ sender: UITextField) {
+        viewModel?.planNameChange(sender.nonNilText, spellingPhase: sender.isSpelling)
+    }
+    
+}
